+++ crates/wdl-analysis/src/handlers.rs
@@ -17,6 +17,7 @@ mod hover;
 mod rename;
 mod semantic_tokens;
 pub(crate) mod snippets;
+mod workspace_symbol;
 
 pub use completions::*;
 pub use document_symbol::*;
@@ -25,6 +26,7 @@ pub use goto_definition::*;
 pub use hover::*;
 pub use rename::*;
 pub use semantic_tokens::*;
+pub use workspace_symbol::*;
 
 /// Context for evaluating expression types during LSP operations.
 ///