+++ crates/wdl-analysis/src/queue.rs
@@ -21,6 +21,7 @@ use lsp_types::GotoDefinitionResponse;
 use lsp_types::Hover;
 use lsp_types::Location;
 use lsp_types::SemanticTokensResult;
+use lsp_types::SymbolInformation;
 use lsp_types::WorkspaceEdit;
 use parking_lot::RwLock;
 use petgraph::Direction;
@@ -80,10 +81,12 @@ pub enum Request<Context> {
     Hover(HoverRequest),
     /// A request to rename a symbol workspace wide.
     Rename(RenameRequest),
-    /// A request to get semantic tokens for a document
+    /// A request to get semantic tokens for a document.
     SemanticTokens(SemanticTokenRequest),
-    /// A request to get symbols for a document
+    /// A request to get symbols for a document.
     DocumentSymbol(DocumentSymbolRequest),
+    /// A request to get symbols for the workspace.
+    WorkspaceSymbol(WorkspaceSymbolRequest),
 }
 
 /// Represents a request to add documents to the graph.
@@ -224,6 +227,14 @@ pub struct DocumentSymbolRequest {
     pub document: Url,
     /// The sender for completing the request.
     pub completed: oneshot::Sender<Option<DocumentSymbolResponse>>,
+}
+
+/// Represents a request to get symbols for the workspace.
+pub struct WorkspaceSymbolRequest {
+    /// The query string to filter symbols.
+    pub query: String,
+    /// The sender for completing the request.
+    pub completed: oneshot::Sender<Option<Vec<SymbolInformation>>>,
 }
 
 /// A simple enumeration to signal a cancellation to the caller.
@@ -633,6 +644,27 @@ where
                         Err(err) => {
                             debug!(
                                 "error occurred while completing document symbol request: {err:?}"
+                            );
+                            completed.send(None).ok();
+                        }
+                    }
+                }
+                Request::WorkspaceSymbol(WorkspaceSymbolRequest { query, completed }) => {
+                    let start = Instant::now();
+                    debug!("received request for workspace symbols with query `{query}`");
+
+                    let graph = self.graph.read();
+                    match handlers::workspace_symbol(&graph, &query) {
+                        Ok(result) => {
+                            debug!(
+                                "workspace symbol request completed in {elapsed:?}",
+                                elapsed = start.elapsed()
+                            );
+                            completed.send(result).ok();
+                        }
+                        Err(err) => {
+                            debug!(
+                                "error occurred while completing workspace symbol request: {err:?}"
                             );
                             completed.send(None).ok();
                         }