+++ crates/wdl-lsp/src/server.rs
@@ -412,6 +412,7 @@ impl LanguageServer for Server {
                     }),
                     ..Default::default()
                 }),
+                workspace_symbol_provider: Some(OneOf::Left(true)),
                 diagnostic_provider: Some(DiagnosticServerCapabilities::Options(
                     DiagnosticOptions {
                         inter_file_dependencies: true,
@@ -932,6 +933,25 @@ impl LanguageServer for Server {
         let result = self
             .analyzer
             .document_symbol(params.text_document.uri)
+            .await
+            .map_err(|e| RpcError {
+                code: ErrorCode::InternalError,
+                message: e.to_string().into(),
+                data: None,
+            })?;
+
+        Ok(result)
+    }
+
+    async fn symbol(
+        &self,
+        params: WorkspaceSymbolParams,
+    ) -> RpcResult<Option<Vec<SymbolInformation>>> {
+        debug!("received `workspace/symbol` request: {params:#?}");
+
+        let result = self
+            .analyzer
+            .workspace_symbol(params.query)
             .await
             .map_err(|e| RpcError {
                 code: ErrorCode::InternalError,