+++ crates/wdl-analysis/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+* Added workspace symbols support for the WDL Language Server ([#588](https://github.com/stjude-rust-labs/wdl/pull/588)).
 * Implemented coercion between `Map` <-> `Object`/`Struct` where the map key
   type <-> `String` ([#586](https://github.com/stjude-rust-labs/wdl/pull/586)).
 * Added snippets support for the WDL Language Server ([#577](https://github.com/stjude-rust-labs/wdl/pull/577)).